requirements.txt
google-generativeai
streamlit
json
python-dotenv